+++ packages/configuration-validation/src/lib.js
@@ -10,6 +10,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
+const merge = require('lodash/merge');
 class ConfigurationValidationError extends Error {}
 
 const configUtil = module.exports;
@@ -28,15 +29,49 @@ const hasDomainTypo = new RegExp('(.com.com)|(://.*){2,}');
 const endsInPath = new RegExp('/$');
 
 configUtil.buildConfigObject = (config) => {
-  return {
-    clientId: config.clientId || config.client_id,
-    issuer: config.issuer,
-    redirectUri: config.redirectUri || config.redirect_uri,
-    tokenManager: {
-      storage: config.storage,
-      autoRenew: config.autoRenew || config.auto_renew
+  // See all supported options: https://github.com/okta/okta-auth-js#configuration-reference
+  // Support for parameters with an underscore will be deprecated in a future release
+  // camelCase was added 2/11/2019: https://github.com/okta/okta-oidc-js/commit/9b04ada6a01c9d9aca391abf0de3e5ecc9811e64
+  
+  config = config || {}; // accept empty
+
+  // Legacy support: allow a property named 'scope' to be either an array or a string.
+  let scopes = config.scopes;
+  if (!scopes && config.scope) {
+    if (Array.isArray(config.scope)) {
+      scopes = config.scope;
+    } else {
+      scopes = config.scope.split(/\s+/);
     }
   }
+
+  // Legacy support: allow TokenManager config 'autoRenew' and 'storage' to be defined at top-level
+  let tokenManager = config.tokenManager;
+  const autoRenew = config.autoRenew || config.auto_renew;
+  const storage = config.storage;
+  if (storage || autoRenew) {
+    // Properties already defined within the "tokenManager" section will not be overwritten
+    tokenManager = merge({
+      autoRenew: autoRenew,
+      storage: storage,
+    }, tokenManager || {});
+  }
+
+  // Legacy support: allow 'responseType' to be a string or an array
+  let responseType = config.responseType || config.response_type;
+  if (typeof responseType === 'string' && responseType.indexOf(' ') >= 0) {
+    responseType = responseType.split(/\s+/);
+  }
+
+  const normalizedConfig = merge({}, config, {
+    clientId: config.clientId || config.client_id,
+    redirectUri: config.redirectUri || config.redirect_uri,
+    responseType: responseType,
+    scopes: scopes,
+    tokenManager: tokenManager,
+  });
+
+  return normalizedConfig;
 }
 
 configUtil.assertIssuer = (issuer, testing = {}) => {
@@ -49,7 +84,6 @@ configUtil.assertIssuer = (issuer, testing = {}) => {
     /* eslint-disable-next-line no-console */
     console.warn(httpsWarning);
   }
-
 
   if (!issuer) {
     throw new ConfigurationValidationError('Your Okta URL is missing. ' + copyMessage);