+++ packages/okta-react-native/android/src/main/java/com/oktareactnative/OktaSdkBridgeModule.java
@@ -15,8 +15,8 @@ package com.oktareactnative;
 import android.app.Activity;
 import android.content.Intent;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.facebook.react.bridge.ActivityEventListener;
 import com.facebook.react.bridge.Arguments;
@@ -27,15 +27,17 @@ import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.WritableMap;
-import com.facebook.react.modules.core.DeviceEventManagerModule;
+import com.facebook.react.modules.core.RCTNativeAppEventEmitter;
 import com.okta.oidc.AuthorizationStatus;
 import com.okta.oidc.OIDCConfig;
 import com.okta.oidc.Okta;
 import com.okta.oidc.RequestCallback;
 import com.okta.oidc.ResultCallback;
 import com.okta.oidc.Tokens;
+import com.okta.oidc.results.Result;
 import com.okta.oidc.clients.sessions.SessionClient;
 import com.okta.oidc.clients.web.WebAuthClient;
+import com.okta.oidc.clients.AuthClient;
 import com.okta.oidc.net.params.TokenTypeHint;
 import com.okta.oidc.net.response.IntrospectInfo;
 import com.okta.oidc.net.response.UserInfo;
@@ -47,6 +49,7 @@ public class OktaSdkBridgeModule extends ReactContextBaseJavaModule implements A
     private final ReactApplicationContext reactContext;
     private OIDCConfig config;
     private WebAuthClient webClient;
+    private AuthClient authClient;
 
     public OktaSdkBridgeModule(ReactApplicationContext reactContext) {
         super(reactContext);
@@ -91,6 +94,13 @@ public class OktaSdkBridgeModule extends ReactContextBaseJavaModule implements A
                     .setRequireHardwareBackedKeyStore(requireHardwareBackedKeyStore)
                     .create();
 
+            this.authClient = new Okta.AuthBuilder()
+                    .withConfig(config)
+                    .withContext(reactContext)
+                    .withStorage(new SharedPreferenceStorage(reactContext))
+                    .setRequireHardwareBackedKeyStore(requireHardwareBackedKeyStore)
+                    .create();
+
             promise.resolve(true);
         } catch (Exception e) {
             promise.reject(OktaSdkError.OKTA_OIDC_ERROR.getErrorCode(), e.getLocalizedMessage(), e);
@@ -118,6 +128,51 @@ public class OktaSdkBridgeModule extends ReactContextBaseJavaModule implements A
         }
 
         webClient.signIn(currentActivity, null);
+    }
+
+    @ReactMethod
+    public void authenticate(String sessionToken) {
+        if (authClient == null) {
+            final WritableMap params = Arguments.createMap();
+            params.putString(OktaSdkConstant.ERROR_CODE_KEY, OktaSdkError.NOT_CONFIGURED.getErrorCode());
+            params.putString(OktaSdkConstant.ERROR_MSG_KEY, OktaSdkError.NOT_CONFIGURED.getErrorMessage());
+            sendEvent(reactContext, OktaSdkConstant.ON_ERROR, params);
+            return;
+        }
+
+        authClient.signIn(sessionToken, null, new RequestCallback<Result, AuthorizationException>() {
+            @Override
+            public void onSuccess(@NonNull Result result) {
+                if (result.isSuccess()) {
+                    try {
+                        SessionClient sessionClient = authClient.getSessionClient();
+                        WritableMap params = Arguments.createMap();
+                        Tokens tokens = sessionClient.getTokens();
+                        params.putString(OktaSdkConstant.RESOLVE_TYPE_KEY, OktaSdkConstant.AUTHORIZED);
+                        params.putString(OktaSdkConstant.ACCESS_TOKEN_KEY, tokens.getAccessToken());
+                        sendEvent(reactContext, OktaSdkConstant.SIGN_IN_SUCCESS, params);
+                    } catch (AuthorizationException e) {
+                        WritableMap params = Arguments.createMap();
+                        params.putString(OktaSdkConstant.ERROR_CODE_KEY, OktaSdkError.SIGN_IN_FAILED.getErrorCode());
+                        params.putString(OktaSdkConstant.ERROR_MSG_KEY, OktaSdkError.SIGN_IN_FAILED.getErrorMessage());
+                        sendEvent(reactContext, OktaSdkConstant.ON_ERROR, params);
+                    }
+                } else {
+                    WritableMap params = Arguments.createMap();
+                    params.putString(OktaSdkConstant.ERROR_CODE_KEY, OktaSdkError.SIGN_IN_FAILED.getErrorCode());
+                    params.putString(OktaSdkConstant.ERROR_MSG_KEY, OktaSdkError.SIGN_IN_FAILED.getErrorMessage());
+                    sendEvent(reactContext, OktaSdkConstant.ON_ERROR, params);
+                }
+            }
+
+            @Override
+            public void onError(String error, AuthorizationException exception) {
+                WritableMap params = Arguments.createMap();
+                params.putString(OktaSdkConstant.ERROR_CODE_KEY, OktaSdkError.OKTA_OIDC_ERROR.getErrorCode());
+                params.putString(OktaSdkConstant.ERROR_MSG_KEY, error);
+                sendEvent(reactContext, OktaSdkConstant.ON_ERROR, params);
+            }
+        });
     }
 
     @ReactMethod
@@ -310,7 +365,7 @@ public class OktaSdkBridgeModule extends ReactContextBaseJavaModule implements A
                            String eventName,
                            @Nullable WritableMap params) {
         reactContext
-                .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                .getJSModule(RCTNativeAppEventEmitter.class)
                 .emit(eventName, params);
     }