+++ packages/okta-react/test/e2e/harness/e2e/page-objects/protected.po.js
@@ -19,7 +19,7 @@ export class ProtectedPage {
   }
 
   waitUntilVisible(path) {
-    Util.waitUrlContains('/protected' + path);
+    Util.waitUrlContains('/protected' + path, 20000);
   }
 
   waitForElement(id) {